Add utils function
